search CVE XMPP Bot added
Simple XMPP bot to query for the last entries in the CVE database
and to get the output in JSON format. Useful for dynamic application
requiring a constant feed of updates.

Current command supported is: last <max>

To get the n last entries from the CVE database.

The sleekxmpp library is required.