Search and show vendor references only in CSV output
Usually you just want to have the references from the vendor to
a specific vulnerability and you don't want the full references.

The -v option allows to search in the hostname of the references
for a specific vendor. Yep, we assume that a vendor got his own
domain name (it's usually the case ;-).

  ./search.py -p webex: -o csv  -v "cisco"

The following example allows you to search webex products and
show only the vendor links.