A ranking functionnality added
Ranking database allows to rank software vulnerabilities based on
their common platform enumeration name. The ranking can be done
per organization or department within your organization or any
meaningful name for you.

As an example, you can add a partial CPE name like "sap:netweaver"
which is very critical for your accounting department.

    ./python3.3 db_ranking.py  -c "sap:netweaver" -g "accounting" -r 3

and then you can lookup the ranking (-r option) for a specific CVE-ID:

    ./python3.3 search.py -c CVE-2012-4341  -r  -n

The ranking can be extended for notification in the XMPP search bot
or alike.

The ranking collection is a new collection in the MongoDB database.

With the following format (a sample record):

   { "_id" : ObjectId("50b2081e597549f61b2a25a7"), "cpe" : "sap:netweaver", "rank" : [ { "accounting" : 3 } ] }