a full-text "search" query added
search is a full-text search on the summary field of each CVE items.
You might consider to have an index on the summary field of the cves collection
, if you expect to have a lot of queries.

search <keyword(s)>

You can search for one or more keywords.
A small help has been also added to the XMPP bot