Crude indexer of CVEs to a Whoosh fulltext index.
The fulltext indexer is relying on Whoosh.

http://packages.python.org/Whoosh/

The indexing is done by enumerating all items from
the MongoDB CVE collection and indexing the summary text of each
CVE. The Path of each document is the CVE-ID.

The current indexing is indexing all CVEs from the MongoDB. This need
to be improved when new items are updated (TODO).

The fulltext indexing is done to overcome the limitation of MongoDB
regarding fulltext indexing and to improve response time for non-indexed
records (especially for the XMPP bot).