Lemmatization and stopwords added to -s in search_fulltext
Initially I used a stemming algorithm to filter the variation of words.
But stemming is not appropriate in this case as the use is not to create
an index out of the terms but to keep the root of the words. Lemmatization is
more appropriate in this case.

The option -s is also checking the stopwords corpus from NTLK to remove
known stopwords from the result.