Added support for vulnerable_product field.
Parser now ingests the field "vulnerable_product" in XML feed from NVD.
New option "--vulnerable-product-only" uses this field. With this
option, "-p" will only return vulnerabilities directly assigned to the
product. I.e. it will not consider "windows_7" if it is only mentioned
as a relevant OS in a "foxit_reader" vulnerability.