Added support for field vulnerable_product
Parser now ingests the field "vulerable_product" from the NVD XML-feed.
New option "--vulnerable-product-only" uses this field:

With this option, "-p" will only return vulnerabilities directly
assigned to the product.

I.e. it will not consider "windows_7" if it is only mentioned as
affected OS in a "foxit_reader" vulnerability.