cleanup: remove obsolete cve_doc.py script
The /api/cve endpoint is incompatible with the public cve.circl.lu API.
While the script still works with the current local API, it relies on
an Asciidoc template that is unlikely to be maintained in sync with
future changes to the CVE collection schema. Removed the inactive tests
related to this script, as well as the igetcve() function from
lib.Query, since the script was its only consumer.